Refactor(Persistence): align identity and domain user lifecycle
Refactor(UserService): Synchronized profile, email, and delete flows with Identity records - to keep user state consistent.
Refactor(AuthenticationService): Replaced DomainUserId-based linkage with shared user IDs - to enforce one-to-one identity mapping.
Refactor(UserRepository): Converted user removal operations to soft deletion - to preserve references while disabling accounts.
Refactor(UserModel): Added soft-delete state and lifecycle fields - to support reversible domain user deactivation.
Refactor(AppDbContext): Enforced shared-key user mapping and explicit refresh token FK configuration - to prevent shadow relations.
Chore(Migration): Generated Identity migration and aligned key types to long across user-related entities - to unblock schema creation.
Chore(API): Standardized configuration on DefaultConnection and TokenSecretKey - to simplify environment-specific setup.
Refactor(URLAnalytics): Updated analytics key types to long - to match URL primary keys during migration generation.

+++ URLShortener.Application/DTOs/EntityDTOs/URL/URLAnalyticsDTO.cs
@@ -2,7 +2,7 @@
 
 public sealed class URLAnalyticsDTO
 {
-    public int ID { get; set; }
+    public long ID { get; set; }
 
     public required string MostUsedLocationsJSON { get; set; }
     public required string MostUsedDevicesJSON { get; set; }
@@ -10,5 +10,5 @@ public sealed class URLAnalyticsDTO
     public int ClickCount { get; set; } = 0;
     public DateTime LastTimeCalculated { get; set; }
 
-    public int URLID { get; set; }
+    public long URLID { get; set; }
 }
